backend/requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.7.2
billiard==4.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
Flask==3.0.2
idna==3.6
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.5
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.3
requests==2.31.0
schedule==1.2.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1